backend/requirements.txt
# PDF Ingestion
PyMuPDF==1.23.7
pdfplumber==0.10.3

# Text Processing and Chunking
transformers==4.40.0
sentence-transformers==2.7.0

# Vector Store
faiss-cpu==1.7.4

# QA and Hugging Face API
torch==2.2.2
accelerate==0.29.1

# Utilities
numpy==1.24.3
scikit-learn==1.3.2
tqdm==4.66.4
requests
python-dotenv

# Optional: FastAPI backend
fastapi==0.111.0
uvicorn==0.30.1

#db
pymongo

#auth 
passlib
python-jose
bcrypt